+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Changes in this release:
 - Extend LsmProject and RemoteServiceInstance to support service creation in alternative initial state. (inmanta-lsm>=5.2.0)
 - Dropped the pkg_resources package
 - Fix bug related to requests CA
+- Added support to retrieve statistics for the orchestrator container spinned up by pytest-inmanta-lsm
 
 # v 4.0.0 (2025-11-19)
 Changes in this release: